Add a data callback ADJ_VEC_GETNORM_CB.
Add a routine adj_storage_set_compare that sets whether adj_record_variable
should compare the value you're currently giving it to one you've previously
recorded. It does this by taking the difference of the two vectors,
and then taking the norm of that difference. It then returns a warning
if the norm is above a (user-given) tolerance.

Test to follow.