Fix a bug with setting up nonlinear blocks. We were doing
c_nblock = NonlinearBlock(...).nblock

where the .nblock refers to the C struct. However! The NonlinearBlock Python
class had a __del__ method that destroyed the C struct .. when the Python
object went out of scope. Which it does immediately. So, it would create
the nonlinear block, assign it, immediately afterwards destroy it, and
then use it.